+++ CHANGELOG.md
@@ -15,7 +15,12 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
+* Unlock the page scroll when the component is destroyed `bBottomSlide`
 * Now styles for dynamically loaded components with `loadModules` are included in the fat-html build `build/monic`
+
+#### :rocket: New Feature
+
+* The destructor, which unlocks the page scroll when the component is destroyed, will be registered once the `lockPageScroll` method is called `iLockPageScroll`
 
 ## v4.0.0-beta.93 (2024-04-19)