+++ src/components/base/b-bottom-slide/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.?? (2024-04-??)
+
+#### :bug: Bug Fix
+
+* Unlock the page scroll when the component is destroyed
+
 ## v4.0.0-beta.93 (2024-04-19)
 
 #### :house: Internal